refactor: use direct imports for uuid, AggregatedError and Scaffold
