+++ packages/client/src/StreamrClient.ts
@@ -3,7 +3,8 @@ import './utils/PatchTsyringe'
 import { container as rootContainer, DependencyContainer } from 'tsyringe'
 
 import { Ethereum } from './Ethereum'
-import { uuid, counterId } from './utils/utils'
+import { uuid } from './utils/uuid'
+import { counterId } from './utils/utils'
 import { pOnce } from './utils/promises'
 import { Debug } from './utils/log'
 import { Context } from './utils/Context'
